requirements.txt
numpy>=1.17.4
pyautogui>=0.9.48
pytesseract>=0.3.0
nptyping>=0.3.1
pillow>=6.2.1
flask>=1.1.1
sounddevice>=0.3.14
soundfile>=0.10.2